move files to respect comp hierarchy
