+++ cmd/agent/subcommands/run/command.go
@@ -45,7 +45,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/api/api/apiimpl"
 	internalAPI "github.com/DataDog/datadog-agent/comp/api/api/def"
 	authtokenimpl "github.com/DataDog/datadog-agent/comp/api/authtoken/createandfetchimpl"
-	commonendpoints "github.com/DataDog/datadog-agent/comp/api/endpoints/common"
+	"github.com/DataDog/datadog-agent/comp/api/commonendpoints"
 	"github.com/DataDog/datadog-agent/comp/collector/collector"
 	"github.com/DataDog/datadog-agent/comp/collector/collector/collectorimpl"
 	"github.com/DataDog/datadog-agent/comp/core"
